+++ lib/ruuuby/class/hsh.rb
@@ -3,4 +3,6 @@
 class ::Hash
   alias_method :🔑?, :key?
   alias_method :∃🔑?, :key?
+  alias_method :∀, :each
+  alias_method :∅?, :empty?
 end