+++ lib/ruuuby/class/nil.rb
@@ -3,4 +3,5 @@
 # add various functions to existing class +NilClass+ (and explicitly create aliases to play nice with IDEs)
 class ::NilClass
   alias_method :empty?, :empty?
+  alias_method :∅?, :empty?
 end