+++ spec/class/hsh_spec.rb
@@ -4,10 +4,7 @@ RSpec.describe 'hsh' do
 
   context 'extends class[Hash]' do
     it 'by creating needed aliases' do
-      expect(::Hash.∃func_alias?(:key?, :🔑?)).to eq(true)
-      expect(::Hash.∃func_alias?(:key?, :∃🔑?)).to eq(true)
-      expect({}.class.∃func_alias?(:key?, :🔑?)).to eq(true)
-      expect({}.class.∃func_alias?(:key?, :∃🔑?)).to eq(true)
+      RuuubyTestHelper::CONFIG_HASH[:aliases].∀{|the_alias, the_func| expect(::Hash.∃func_alias?(the_func, the_alias)).to eq(true)}
     end
   end