+++ spec/class/str_spec.rb
@@ -2,20 +2,15 @@
 
 RSpec.describe 'str' do
   let(:data_empty){''}
-  let(:wrong_param_type){Ruuuby::Err::WrongParamType}
+  let(:wrong_param_type){Ruuuby::ParamErr::WrongParamType}
 
   context 'extends class[String]' do
 
     it 'by creating needed aliases' do
-      expect(::String.∃func?(:∋?)).to eq(true)
-      expect(::String.∃func?(:∈?)).to eq(true)
-      expect(::String.∃func?(:∌?)).to eq(true)
-      expect(::String.∃func?(:∉?)).to eq(true)
-      expect(::String.∃func?(:>>)).to eq(true)
-      expect(::String.∃func?(:ensure_ending!)).to eq(true)
-      expect(::String.∃func?(:ensure_start!)).to eq(true)
+      RuuubyTestHelper::CONFIG_STRING[:ruby].∀{ |func| expect(::String.∃func?(func)).to eq(true) }
+      RuuubyTestHelper::CONFIG_STRING[:c].∀{ |func| expect(::String.∃func?(func)).to eq(true) }
+      RuuubyTestHelper::CONFIG_STRING[:aliases].∀{ |aliased_func, base_func| expect(::String.∃func?(aliased_func)).to eq(true) }
     end
-
 
     context 'func{ensure_ending!}' do
       context 'handles cases' do
@@ -24,46 +19,25 @@ RSpec.describe 'str' do
             it 'passes simple scenarios' do
               [
                   ['', '', ''], ['', 'aaa', 'aaa'], ['aaa', '', 'aaa'], %w(hello? ? hello?), %w(hello ?a hello?a), %w(hellb ?? hellb??), %w(hello? ?? hello??)
-              ].∀{|a|expect(a[0].ensure_ending!(a[1], true)).to eq(a[2])}
+              ].∀{|a|expect(a[0].ensure_ending!(a[1])).to eq(a[2])}
             end
             it 'passes complex scenarios' do
               [
                   %w(baa aaa baaa), %w(bba aaa bbaaa), %w(baa aaaaaaaaa baaaaaaaaa),
                   %w(baaaaa aaaaaa baaaaaa), %w(ba aaaaaaaa baaaaaaaa), %w(abc bca abca),
                   ['hi ', ' ', 'hi '], [' hi', ' ', ' hi '], %w(abc bcd abcd)
-              ].∀{|a|expect(a[0].ensure_ending!(a[1], true)).to eq(a[2])}
+              ].∀{|a|expect(a[0].ensure_ending!(a[1])).to eq(a[2])}
             end
           end # end context 'positive' -> 'with partial fill in'
-          context 'without partial fill in' do
-            it 'passes simple scenarios' do
-              [
-                  ['', '', ''], ['', 'aaa', 'aaa'], ['aaa', '', 'aaa'], %w(hello? ? hello?), %w(hello ?a hello?a), %w(hellb ?? hellb??), %w(hello? ?? hello???)
-              ].∀{|a|expect(a[0].ensure_ending!(a[1], false)).to eq(a[2])}
-            end
-            it 'passes complex scenarios' do
-              [
-                  %w(baa aaa baaaaa), %w(bba aaa bbaaaa), %w(baa aaaaaaaaa baaaaaaaaaaa),
-                  %w(baaaaa aaaaaa baaaaaaaaaaa), %w(ba aaaaaaaa baaaaaaaaa), %w(abc bca abcbca),
-                  ['hi ', ' ', 'hi '], [' hi', ' ', ' hi '], %w(abc bcd abcbcd)
-              ].∀{|a|expect(a[0].ensure_ending!(a[1], false)).to eq(a[2])}
-            end
-          end # end context 'positive' -> 'without partial fill in'
         end # end context 'positive'
         context 'error' do
           it 'catches bad param: start' do
-            [nil, 1337, {}].∀{|a|expect{''.ensure_ending!(a, true)}.to raise_exception(ArgumentError)}
-            [nil, 1337, {}].∀{|a|expect{''.ensure_ending!(a, false)}.to raise_exception(ArgumentError)}
-          end
-          it 'catches bad param: use_partial_fill_in' do
-            expect{''.ensure_ending!('nil', nil)}.to raise_exception(ArgumentError)
-            expect{''.ensure_ending!('1337', 1337)}.to raise_exception(ArgumentError)
-            expect{''.ensure_ending!('', {})}.to raise_exception(ArgumentError)
+            [nil, 1337, {}].∀{|a|expect{''.ensure_ending!(a)}.to raise_exception(ArgumentError)}
           end
           it 'catches frozen string' do
             a = 'my_frozen_string'.❄️
             a.❄️
-            expect{a.ensure_ending!('bbb', true)}.to raise_error(FrozenError)
-            expect{a.ensure_ending!('bbb', false)}.to raise_error(FrozenError)
+            expect{a.ensure_ending!('bbb')}.to raise_error(FrozenError)
           end
         end # end context 'error'
       end
@@ -80,32 +54,66 @@ RSpec.describe 'str' do
     end
 
     context 'func{∌?} (excluded?)' do
+      context 'handles needed scenarios' do
+        it 'cases: positive' do
+          expect('abc'.∌? 'd').to eq(true)
+        end
+        it 'cases: negative' do
+          expect('abc'.∌? 'b').to eq(false)
+        end
+        it 'cases: error' do
+          expect{'b'.∌? nil}.to raise_exception(ArgumentError)
+          expect{'b'.∌? 1337}.to raise_exception(ArgumentError)
+          expect{'b'.∌? %w(a cc b)}.to raise_exception(ArgumentError)
+        end
+      end
       it 'works correctly' do
         expect('abc'.∌? 'd').to eq(true)
         expect('abc'.∌? 'b').to eq(false)
       end
-      it 'catches bad arg' do
-        expect{'b'.∌? nil}.to raise_exception(ArgumentError)
-      end
     end
 
     context 'func{∈?} (include?)' do
-      it 'works correctly' do
-        expect('b'.∈? 'abc').to eq(true)
-        expect('d'.∈? 'abc').to eq(false)
+      context 'handles needed scenarios' do
+        it 'cases: positive' do
+          expect('b'.∈? 'abc').to eq(true)
+          expect('b'.∈? %w(a cc b)).to eq(true)
+        end
+        it 'cases: negative' do
+          expect('d'.∈? 'abc').to eq(false)
+        end
+        it 'cases: error' do
+          expect{'b'.∈? nil}.to raise_exception(ArgumentError)
+          expect{'b'.∈? 1337}.to raise_exception(ArgumentError)
+        end
       end
-      it 'catches bad arg' do
-        expect{'b'.∈? nil}.to raise_exception(ArgumentError)
+    end
+
+    context 'func{∅?} (alias for "remove_empty!")' do
+      context 'handles needed scenarios' do
+        it 'cases: positive' do
+          expect(''.∅?).to eq(true)
+        end
+        it 'cases: negative' do
+          expect(' '.∅?).to eq(false)
+        end
       end
     end
 
     context 'func{∉?} (excluded?)' do
-      it 'works correctly' do
-        expect('d'.∉? 'abc').to eq(true)
-        expect('b'.∉? 'abc').to eq(false)
-      end
-      it 'catches bad arg' do
-        expect{'b'.∉? nil}.to raise_exception(ArgumentError)
+      context 'handles needed scenarios' do
+        it 'cases: positive' do
+          expect('d'.∉? 'abc').to eq(true)
+          expect('d'.∉? %w(a cc b)).to eq(true)
+        end
+        it 'cases: negative' do
+          expect('b'.∉? 'abc').to eq(false)
+          expect('b'.∉? %w(a cc b)).to eq(false)
+        end
+        it 'cases: error' do
+          expect{'b'.∉? nil}.to raise_exception(ArgumentError)
+          expect{'b'.∉? 1337}.to raise_exception(ArgumentError)
+        end
       end
     end
 
@@ -152,45 +160,25 @@ RSpec.describe 'str' do
             it 'simple data' do
               [
                   ['', '', ''], ['', ' ', ' '], [' ', ' ', ' '], ['', 'a', 'a'], %w(b a ab), %w(c aaac aaac)
-              ].∀{|a|expect(a[0].ensure_start!(a[1], true)).to eq(a[2])}
+              ].∀{|a|expect(a[0].ensure_start!(a[1])).to eq(a[2])}
             end
             it 'complex data' do
               [
                   %w(baa b baa), %w(baa ba baa), %w(baa baa baa), %w(baa bb bbaa), %w(baa bba bbaa),
                   %w(baa bbaa bbaa), %w(baa bbaaa bbaaabaa), %w(abc123xyz 123 123abc123xyz)
-              ].∀{|a|expect(a[0].ensure_start!(a[1], true)).to eq(a[2])}
-            end
-          end
-          context 'without partial fill in' do
-            it 'simple data' do
-              [
-                  ['', '', ''], ['', ' ', ' '], [' ', ' ', ' '], ['', 'a', 'a'], %w(b a ab), %w(b aaab aaabb)
-              ].∀{|a|expect(a[0].ensure_start!(a[1], false)).to eq(a[2])}
-            end
-            it 'complex data' do
-              [
-                  %w(baa b baa), %w(baa ba baa), %w(baa baa baa), %w(baa bb bbbaa), %w(baa bba bbabaa),
-                  %w(baa bbaa bbaabaa), %w(baa bbaaa bbaaabaa), %w(abc123xyz 123 123abc123xyz)
-              ].∀{|a|expect(a[0].ensure_start!(a[1], false)).to eq(a[2])}
+              ].∀{|a|expect(a[0].ensure_start!(a[1])).to eq(a[2])}
             end
           end
         end # end context 'positive'
 
         context 'error' do
           it 'catches bad param: start' do
-            [nil, 1337, {}].∀{|a|expect{''.ensure_start!(a, true)}.to raise_exception(ArgumentError)}
-            [nil, 1337, {}].∀{|a|expect{''.ensure_start!(a, false)}.to raise_exception(ArgumentError)}
-          end
-          it 'catches bad param: use_partial_fill_in' do
-            expect{''.ensure_start!('nil', nil)}.to raise_exception(ArgumentError)
-            expect{''.ensure_start!('1337', 1337)}.to raise_exception(ArgumentError)
-            expect{''.ensure_start!('', {})}.to raise_exception(ArgumentError)
+            [nil, 1337, {}].∀{|a|expect{''.ensure_start!(a)}.to raise_exception(ArgumentError)}
           end
           it 'catches frozen string' do
             a = 'my_frozen_string'.❄️
             a.❄️
-            expect{a.ensure_start!('bbb', true)}.to raise_error(FrozenError)
-            expect{a.ensure_start!('bbb', false)}.to raise_error(FrozenError)
+            expect{a.ensure_start!('bbb')}.to raise_error(FrozenError)
           end
         end
       end
@@ -202,41 +190,45 @@ RSpec.describe 'str' do
   # |    |___ |  \ |    \__/ |  \  |  | /~~\ | \| \__, |___
   context 'performance', :'performance' do
     let(:big_str){'ABT$JM^#$^ABT$JM^#$^ABT$JM^#$^ABT$JM^#$^ABT$JM^#$^ABT$JM^#$^ABT$JM^#$^ABT$JM^#$^ABT$JM^#$^ABT$JM^#$^ABT$JM^#$^ABT$JM^#$^'}
+    let(:a_str){'any54wyv45hv'}
 
     it 'func[∋?] runs fast enough' do
-      expect{'any54wyv45hv'.∋? 'c'}.to perform_very_quickly
+      expect{a_str.∋? 'c'}.to perform_very_quickly
     end
 
     it 'func[∌?] runs fast enough' do
-      expect{'any54wyv45hv'.∌? 'c'}.to perform_very_quickly
+      expect{a_str.∌? 'c'}.to perform_very_quickly
     end
 
     it 'func[∈?] runs fast enough' do
-      expect{'any54wyv45hv'.∈? 'c'}.to perform_very_quickly
+      expect{a_str.∈? 'c'}.to perform_very_quickly
     end
 
     it 'func[∉?] runs fast enough' do
-      expect{'any54wyv45hv'.∉? 'c'}.to perform_very_quickly
+      expect{a_str.∉? 'c'}.to perform_very_quickly
     end
 
     it 'func[>>] runs fast enough' do
-      expect{'any54wyv45hv'.>> 'bASDVASb5t4t'}.to perform_very_quickly
+      expect{a_str.>> 'bASDVASb5t4t'}.to perform_very_quickly
     end
 
     context 'with partial fill in, performs quickly' do
       it 'func[ensure_ending!]' do
-        [%w(hello ?a), ['', big_str], [big_str, '']].∀{|a|expect{a[0].ensure_ending!(a[1], true)}.to perform_quickly}
+        [%w(hello ?a), ['', big_str], [big_str, '']].∀{|a|expect{a[0].ensure_ending!(a[1])}.to perform_quickly}
       end
       it 'func[ensure_start!]' do
-        [%w(hello ?a), ['', big_str], [big_str, '']].∀{|a|expect{a[0].ensure_start!(a[1], true)}.to perform_quickly}
+        [%w(hello ?a), ['', big_str], [big_str, '']].∀{|a|expect{a[0].ensure_start!(a[1])}.to perform_quickly}
       end
     end
-    context 'without partial fill in, performs quickly' do
-      it 'func[ensure_ending!]' do
-        [%w(hello ?a), ['', big_str], [big_str, '']].∀{|a|expect{a[0].ensure_ending!(a[1], false)}.to perform_quickly}
-      end
-      it 'func[ensure_start!]' do
-        [%w(hello ?a), ['', big_str], [big_str, '']].∀{|a|expect{a[0].ensure_start!(a[1], false)}.to perform_quickly}
+
+    context 'func{∅?} (alias for "remove_empty!")' do
+      context 'handles needed scenarios extremely quickly' do
+        it 'cases: positive' do
+          expect{''.∅?}.to perform_very_quickly
+        end
+        it 'cases: negative' do
+          expect{' '.∅?}.to perform_very_quickly
+        end
       end
     end
   end