+++ spec/helpers/static_test_data.rb
@@ -1,0 +1,9 @@
+
+class TestDataGlobalClass
+  class TestDataInternalClass
+    class TestDataInternalNestedClass
+    end
+  end
+  module TestDataInternalModule
+  end
+end